Separate announcement code out from web parsing code. Add intermediate layer between parsing and specific announcement method.
