+++ ses/speech.py
@@ -20,3 +20,28 @@ def sayText(sentence):
         # Works just fine after the correct libraries are installed
         speaker = win32com.client.Dispatch("SAPI.SpVoice")
         speaker.Speak(sentence)
+
+def announce(job):
+    # Generate sentence to speak
+    words = []
+    words.append('There is an incident requiring ')
+    words.append(job['priority'])
+    words.append(' response! ')
+
+    words.append(' It is a ')
+    words.append(job['type'])
+    words.append(' !')
+
+    words.append(' Location is ')
+    words.append(job['address'])
+    words.append(' !')
+
+
+    # Say it!
+    announcement = ' '.join(words)
+    print(announcement)
+    # Repeat announcement twice so they can be heard a bit clearly
+    sayText(announcement)
+    sayText("I repeat,")
+    sayText(announcement)
+    #sayText('S E S, Rollout!')
